+++ Common/src/main/java/fuzs/mindfuldarkness/client/gui/components/NewTextureButton.java
@@ -2,7 +2,7 @@ package fuzs.mindfuldarkness.client.gui.components;
 
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.PoseStack;
-import fuzs.mindfuldarkness.client.handler.DaytimeSwitchHandler;
+import fuzs.mindfuldarkness.client.handler.DaytimeSwitcherHandler;
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.Font;
@@ -28,7 +28,7 @@ public class NewTextureButton extends Button {
         Minecraft minecraft = Minecraft.getInstance();
         Font font = minecraft.font;
         RenderSystem.setShader(GameRenderer::getPositionTexShader);
-        RenderSystem.setShaderTexture(0, DaytimeSwitchHandler.TEXTURE_LOCATION);
+        RenderSystem.setShaderTexture(0, DaytimeSwitcherHandler.TEXTURE_LOCATION);
         RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, this.alpha);
         int i = this.getYImage(this.isHoveredOrFocused());
         RenderSystem.enableBlend();