+++ Fabric/src/main/java/fuzs/mindfuldarkness/client/MindfulDarknessFabricClient.java
@@ -1,10 +1,9 @@
 package fuzs.mindfuldarkness.client;
 
-import fuzs.mindfuldarkness.MindfulDarkness;
 import fuzs.mindfuldarkness.api.client.event.ContainerScreenEvents;
 import fuzs.mindfuldarkness.api.client.event.ExtraScreenEvents;
-import fuzs.mindfuldarkness.client.handler.DaytimeSwitchHandler;
-import fuzs.puzzleslib.client.core.ClientFactories;
+import fuzs.mindfuldarkness.client.handler.DaytimeSwitcherHandler;
+import fuzs.mindfuldarkness.client.handler.DaytimeMenuHandler;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.screen.v1.ScreenEvents;
@@ -18,14 +17,18 @@ public class MindfulDarknessFabricClient implements ClientModInitializer {
     }
 
     private static void registerHandlers() {
-        ContainerScreenEvents.BACKGROUND.register(DaytimeSwitchHandler::onContainerScreen$Render$Background);
+        ContainerScreenEvents.BACKGROUND.register(DaytimeSwitcherHandler::onContainerScreen$Render$Background);
         ScreenEvents.AFTER_INIT.register((client, screen, scaledWidth, scaledHeight) -> {
-            DaytimeSwitchHandler.onScreenInit$Post(screen, client, scaledWidth, scaledHeight, abstractWidget -> {
+            DaytimeSwitcherHandler.onScreenInit$Post(screen, client, scaledWidth, scaledHeight, abstractWidget -> {
+                Screens.getButtons(screen).add(abstractWidget);
+                return abstractWidget;
+            });
+            DaytimeMenuHandler.onScreenInit$Post(screen, client, scaledWidth, scaledHeight, abstractWidget -> {
                 Screens.getButtons(screen).add(abstractWidget);
                 return abstractWidget;
             });
         });
-        ExtraScreenEvents.OPENING.register(DaytimeSwitchHandler::onScreenOpen);
-        ClientTickEvents.END_CLIENT_TICK.register(DaytimeSwitchHandler::onClientTick$End);
+        ExtraScreenEvents.OPENING.register(DaytimeSwitcherHandler::onScreenOpen);
+        ClientTickEvents.END_CLIENT_TICK.register(DaytimeSwitcherHandler::onClientTick$End);
     }
 }