+++ Forge/src/main/java/fuzs/mindfuldarkness/client/MindfulDarknessForgeClient.java
@@ -1,8 +1,8 @@
 package fuzs.mindfuldarkness.client;
 
 import fuzs.mindfuldarkness.MindfulDarkness;
-import fuzs.mindfuldarkness.client.handler.DaytimeSwitchHandler;
-import fuzs.puzzleslib.client.core.ClientFactories;
+import fuzs.mindfuldarkness.client.handler.DaytimeSwitcherHandler;
+import fuzs.mindfuldarkness.client.handler.DaytimeMenuHandler;
 import net.minecraft.client.Minecraft;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.client.event.ContainerScreenEvent;
@@ -23,16 +23,20 @@ public class MindfulDarknessForgeClient {
 
     private static void registerHandlers() {
         MinecraftForge.EVENT_BUS.addListener((final ContainerScreenEvent.Render.Background evt) -> {
-            DaytimeSwitchHandler.onContainerScreen$Render$Background(evt.getContainerScreen(), evt.getPoseStack(), evt.getMouseX(), evt.getMouseY());
+            DaytimeSwitcherHandler.onContainerScreen$Render$Background(evt.getContainerScreen(), evt.getPoseStack(), evt.getMouseX(), evt.getMouseY());
         });
         MinecraftForge.EVENT_BUS.addListener((final ScreenEvent.Init.Post evt) -> {
-            DaytimeSwitchHandler.onScreenInit$Post(evt.getScreen(), Minecraft.getInstance(), evt.getScreen().width, evt.getScreen().height, abstractWidget -> {
+            DaytimeSwitcherHandler.onScreenInit$Post(evt.getScreen(), Minecraft.getInstance(), evt.getScreen().width, evt.getScreen().height, abstractWidget -> {
+                evt.addListener(abstractWidget);
+                return abstractWidget;
+            });
+            DaytimeMenuHandler.onScreenInit$Post(evt.getScreen(), Minecraft.getInstance(), evt.getScreen().width, evt.getScreen().height, abstractWidget -> {
                 evt.addListener(abstractWidget);
                 return abstractWidget;
             });
         });
         MinecraftForge.EVENT_BUS.addListener((final ScreenEvent.Opening evt) -> {
-            DaytimeSwitchHandler.onScreenOpen(evt.getCurrentScreen(), evt.getNewScreen()).ifPresent(newScreen -> {
+            DaytimeSwitcherHandler.onScreenOpen(evt.getCurrentScreen(), evt.getNewScreen()).ifPresent(newScreen -> {
                 if (newScreen == evt.getCurrentScreen()) {
                     evt.setCanceled(true);
                 } else {
@@ -41,7 +45,7 @@ public class MindfulDarknessForgeClient {
             });
         });
         MinecraftForge.EVENT_BUS.addListener((final TickEvent.ClientTickEvent evt) -> {
-            if (evt.phase == TickEvent.Phase.END) DaytimeSwitchHandler.onClientTick$End(Minecraft.getInstance());
+            if (evt.phase == TickEvent.Phase.END) DaytimeSwitcherHandler.onClientTick$End(Minecraft.getInstance());
         });
     }
 }